requirements.txt
discord.py
requests~=2.25.1

beautifulsoup4
#lxml
#beautifulsoup4~=4.9.3
#bs4~=0.0.1